Add Coordinator and fix issues
In this commit I added coordinator in order to decouple the Feature SPM's to be as much independent as possible.
I also found an issue, onTapGesture inside the Article Cell looks like is firing incorrect article details if user touch at right half of the cell. Fixed with rectangle content shape.
This is because the Lazy Grid and this contentShape defines the tappable area